requirements.txt
streamlit
torch
torchvision
scikit-learn
numpy
pandas
plotly
Pillow
matplotlib
